// src/containers/budget/enums.ts
export enum ExpensesActionTypes {
	SET_EXPENSES = 'SET_EXPENSES',
	SET_EXPENSES_TABLE = 'SET_EXPENSES_TABLE',
	DELETE_CATEGORY_COLUMN = 'DELETE_CATEGORY_COLUMN',
	SET_TASK_TITLE = 'SET_TASK_TITLE',
	SET_EXPENSE_TITLE = 'SET_EXPENSE_TITLE',
	REMOVE_EXPENSE_LIST='REMOVE_EXPENSE_LIST',
	ADD_EXPENSE_LIST= 'ADD_EXPENSE_LIST',
	UPDATE_EXPENSE_LIST= 'UPDATE_EXPENSE_LIST',
	ADD_EXPENSES_TABLE = 'ADD_EXPENSES_TABLE',
	ADD_ROW_TO_EXPENSE_TABLE = 'ADD_ROW_TO_EXPENSE_TABLE',
	SET_PAYMENTS = 'SET_PAYMENTS',
	DELETE_PAYMENTS = 'DELETE_PAYMENTS'
}

export enum IncomeActionTypes {
	SET_INCOME = 'SET_INCOME'
}